sbi-life-portal-master/sbi-life-portal-master/pyproject.toml
[tool.poetry]
name = "sbi-life"
version = "0.1.0"
description = "Claim Application Processing"
authors = ["Mohit Shah <mohit@glib.ai>"]

[tool.poetry.dependencies]
python = "3.10.13"
Django = "4.1.9"
django-environ = "^0.8.1"
django-debug-toolbar = "^4.2.0"
django-tables2 = "^2.4.1"
django-allauth = "^0.47.0"
django-filter = "^21.1"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.2"
django-widget-tweaks = "^1.5.0"
python-magic = "^0.4.25"
PyPDF2 = "^2.10.3"
dexter-sbi-life-processor = { git = "https://github.com/GlibAI/dexter-sbi-life-processor.git", rev="0.0.40" }
celery = "^5.2.1"
django-celery-results = "^2.3.1"
django-celery-beat = "^2.5.0"
django-redis = "^5.4.0"
django-storages = "^1.12.3"
XlsxWriter = "^3.0.2"
openpyxl = "^3.0.9"
boto3 = "^1.20.26"
django-cors-headers = "^4.3.0"
django-auditlog = "^2.3.0"
pysftp = "^0.2.9"
weasyprint = "^60.1"
django-ses = "^3.5.0"
coverage = "^7.4.0"
